# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dashboard Data Desa API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface DesaSummary {
  totalDesa: number;
  totalPenduduk: number;
  totalKK: number;
  totalLuasHa: number;
  totalDanaDesa2024: number;
  rataRealisasiPct: number;
  rataSkorIDM: number;
  desaMandiri: number;
  desaMaju: number;
  desaBerkembang: number;
  desaTertinggal: number;
  desaSangatTertinggal: number;
  kecamatan: string;
  kabupaten: string;
  provinsi: string;
  lastUpdate: string;
}

export interface ProfilDesa {
  id: number;
  nama: string;
  kecamatan: string;
  kabupaten: string;
  luas_ha: number;
  jumlah_penduduk: number;
  jumlah_kk: number;
  kepala_desa: string;
  jenis_desa: string;
  lat: number;
  lng: number;
  tahun_berdiri: number;
}

export interface ProfilDesaInput {
  nama: string;
  kecamatan: string;
  kabupaten: string;
  luas_ha: number;
  jumlah_penduduk: number;
  jumlah_kk: number;
  kepala_desa: string;
  jenis_desa: string;
  lat: number;
  lng: number;
  tahun_berdiri: number;
}

export interface DanaDesa {
  id: number;
  desa_id: number;
  nama_desa: string;
  tahun: number;
  alokasi: number;
  realisasi: number;
  persen_realisasi: number;
  infra_pct: number;
  pemberdayaan_pct: number;
  operasional_pct: number;
  sumber: string;
}

export interface IDMRecord {
  id: number;
  desa_id: number;
  nama_desa: string;
  tahun: number;
  skor_idm: number;
  status_idm: string;
  iks: number;
  ike: number;
  ikl: number;
  skor_sebelumnya: number;
  delta: number;
}

export interface PelayananDesa {
  id: number;
  nama_desa: string;
  jenis: string;
  nama_fasilitas: string;
  status: string;
  pengguna_bulan: number;
  keterangan: string;
}

export interface PelayananDesaInput {
  nama_desa: string;
  jenis: string;
  nama_fasilitas: string;
  status: string;
  pengguna_bulan: number;
  keterangan: string;
}