# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Dashboard Data Desa API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: desa
    description: Village data operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /desa/summary:
    get:
      operationId: getDesaSummary
      tags: [desa]
      summary: Get dashboard summary KPIs
      responses:
        "200":
          description: Summary KPIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesaSummary"
    put:
      operationId: updateDesaSummary
      tags: [desa]
      summary: Update dashboard summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesaSummary"
      responses:
        "200":
          description: Updated summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesaSummary"

  /desa/profil:
    get:
      operationId: getDesaProfil
      tags: [desa]
      summary: Get all village profiles
      responses:
        "200":
          description: List of village profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfilDesa"
    post:
      operationId: createDesaProfil
      tags: [desa]
      summary: Create a village profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfilDesaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilDesa"

  /desa/profil/{id}:
    put:
      operationId: updateDesaProfil
      tags: [desa]
      summary: Update a village profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfilDesaInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilDesa"
    delete:
      operationId: deleteDesaProfil
      tags: [desa]
      summary: Delete a village profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /desa/dana:
    get:
      operationId: getDesaDana
      tags: [desa]
      summary: Get dana desa records
      responses:
        "200":
          description: List of dana desa records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DanaDesa"
    post:
      operationId: createDesaDana
      tags: [desa]
      summary: Create a dana desa record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DanaDesa"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DanaDesa"

  /desa/dana/{id}:
    put:
      operationId: updateDesaDana
      tags: [desa]
      summary: Update a dana desa record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DanaDesa"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DanaDesa"
    delete:
      operationId: deleteDesaDana
      tags: [desa]
      summary: Delete a dana desa record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /desa/idm:
    get:
      operationId: getDesaIDM
      tags: [desa]
      summary: Get IDM records
      responses:
        "200":
          description: List of IDM records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IDMRecord"
    post:
      operationId: createDesaIDM
      tags: [desa]
      summary: Create an IDM record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IDMRecord"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IDMRecord"

  /desa/idm/{id}:
    put:
      operationId: updateDesaIDM
      tags: [desa]
      summary: Update an IDM record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IDMRecord"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IDMRecord"
    delete:
      operationId: deleteDesaIDM
      tags: [desa]
      summary: Delete an IDM record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /desa/pelayanan:
    get:
      operationId: getDesaPelayanan
      tags: [desa]
      summary: Get village services data
      responses:
        "200":
          description: List of village services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PelayananDesa"
    post:
      operationId: createDesaPelayanan
      tags: [desa]
      summary: Create a village service record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PelayananDesaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PelayananDesa"

  /desa/pelayanan/{id}:
    put:
      operationId: updateDesaPelayanan
      tags: [desa]
      summary: Update a village service record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PelayananDesaInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PelayananDesa"
    delete:
      operationId: deleteDesaPelayanan
      tags: [desa]
      summary: Delete a village service record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    DesaSummary:
      type: object
      properties:
        totalDesa: { type: integer }
        totalPenduduk: { type: integer }
        totalKK: { type: integer }
        totalLuasHa: { type: number }
        totalDanaDesa2024: { type: number }
        rataRealisasiPct: { type: number }
        rataSkorIDM: { type: number }
        desaMandiri: { type: integer }
        desaMaju: { type: integer }
        desaBerkembang: { type: integer }
        desaTertinggal: { type: integer }
        desaSangatTertinggal: { type: integer }
        kecamatan: { type: string }
        kabupaten: { type: string }
        provinsi: { type: string }
        lastUpdate: { type: string }
      required: [totalDesa, totalPenduduk, totalKK, totalLuasHa, totalDanaDesa2024, rataRealisasiPct, rataSkorIDM, desaMandiri, desaMaju, desaBerkembang, desaTertinggal, desaSangatTertinggal, kecamatan, kabupaten, provinsi, lastUpdate]

    ProfilDesa:
      type: object
      required: [id, nama, kecamatan, kabupaten, luas_ha, jumlah_penduduk, jumlah_kk, kepala_desa, jenis_desa, lat, lng, tahun_berdiri]
      properties:
        id: { type: integer }
        nama: { type: string }
        kecamatan: { type: string }
        kabupaten: { type: string }
        luas_ha: { type: number }
        jumlah_penduduk: { type: integer }
        jumlah_kk: { type: integer }
        kepala_desa: { type: string }
        jenis_desa: { type: string }
        lat: { type: number }
        lng: { type: number }
        tahun_berdiri: { type: integer }

    ProfilDesaInput:
      type: object
      required: [nama, kecamatan, kabupaten, luas_ha, jumlah_penduduk, jumlah_kk, kepala_desa, jenis_desa, lat, lng, tahun_berdiri]
      properties:
        nama: { type: string }
        kecamatan: { type: string }
        kabupaten: { type: string }
        luas_ha: { type: number }
        jumlah_penduduk: { type: integer }
        jumlah_kk: { type: integer }
        kepala_desa: { type: string }
        jenis_desa: { type: string }
        lat: { type: number }
        lng: { type: number }
        tahun_berdiri: { type: integer }

    DanaDesa:
      type: object
      required: [id, desa_id, nama_desa, tahun, alokasi, realisasi, persen_realisasi, infra_pct, pemberdayaan_pct, operasional_pct, sumber]
      properties:
        id: { type: integer }
        desa_id: { type: integer }
        nama_desa: { type: string }
        tahun: { type: integer }
        alokasi: { type: number }
        realisasi: { type: number }
        persen_realisasi: { type: number }
        infra_pct: { type: number }
        pemberdayaan_pct: { type: number }
        operasional_pct: { type: number }
        sumber: { type: string }

    IDMRecord:
      type: object
      required: [id, desa_id, nama_desa, tahun, skor_idm, status_idm, iks, ike, ikl, skor_sebelumnya, delta]
      properties:
        id: { type: integer }
        desa_id: { type: integer }
        nama_desa: { type: string }
        tahun: { type: integer }
        skor_idm: { type: number }
        status_idm: { type: string }
        iks: { type: number }
        ike: { type: number }
        ikl: { type: number }
        skor_sebelumnya: { type: number }
        delta: { type: number }

    PelayananDesa:
      type: object
      required: [id, nama_desa, jenis, nama_fasilitas, status, pengguna_bulan, keterangan]
      properties:
        id: { type: integer }
        nama_desa: { type: string }
        jenis: { type: string }
        nama_fasilitas: { type: string }
        status: { type: string }
        pengguna_bulan: { type: integer }
        keterangan: { type: string }

    PelayananDesaInput:
      type: object
      required: [nama_desa, jenis, nama_fasilitas, status, pengguna_bulan, keterangan]
      properties:
        nama_desa: { type: string }
        jenis: { type: string }
        nama_fasilitas: { type: string }
        status: { type: string }
        pengguna_bulan: { type: integer }
        keterangan: { type: string }