# lib/api-zod/src/generated/types/danaDesa.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dashboard Data Desa API
 * OpenAPI spec version: 0.2.0
 */

export interface DanaDesa {
  id: number;
  desa_id: number;
  nama_desa: string;
  tahun: number;
  alokasi: number;
  realisasi: number;
  persen_realisasi: number;
  infra_pct: number;
  pemberdayaan_pct: number;
  operasional_pct: number;
  sumber: string;
}