# lib/api-zod/src/generated/types/desaSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dashboard Data Desa API
 * OpenAPI spec version: 0.2.0
 */

export interface DesaSummary {
  totalDesa: number;
  totalPenduduk: number;
  totalKK: number;
  totalLuasHa: number;
  totalDanaDesa2024: number;
  rataRealisasiPct: number;
  rataSkorIDM: number;
  desaMandiri: number;
  desaMaju: number;
  desaBerkembang: number;
  desaTertinggal: number;
  desaSangatTertinggal: number;
  kecamatan: string;
  kabupaten: string;
  provinsi: string;
  lastUpdate: string;
}