# lib/api-zod/src/generated/types/iDMRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dashboard Data Desa API
 * OpenAPI spec version: 0.2.0
 */

export interface IDMRecord {
  id: number;
  desa_id: number;
  nama_desa: string;
  tahun: number;
  skor_idm: number;
  status_idm: string;
  iks: number;
  ike: number;
  ikl: number;
  skor_sebelumnya: number;
  delta: number;
}