# lib/api-zod/src/generated/types/pelayananDesaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dashboard Data Desa API
 * OpenAPI spec version: 0.2.0
 */

export interface PelayananDesaInput {
  nama_desa: string;
  jenis: string;
  nama_fasilitas: string;
  status: string;
  pengguna_bulan: number;
  keterangan: string;
}