# lib/api-zod/src/generated/types/profilDesa.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dashboard Data Desa API
 * OpenAPI spec version: 0.2.0
 */

export interface ProfilDesa {
  id: number;
  nama: string;
  kecamatan: string;
  kabupaten: string;
  luas_ha: number;
  jumlah_penduduk: number;
  jumlah_kk: number;
  kepala_desa: string;
  jenis_desa: string;
  lat: number;
  lng: number;
  tahun_berdiri: number;
}